requirements.txt
# Core data science and modeling libraries
pandas
scikit-learn
xgboost
imbalanced-learn
joblib

# Libraries for Azure interaction
azure-identity
azure-storage-blob
azure-ai-ml